Frequency of genotypes for two-point recombination fraction estimation

Returns the frequency of each genotype for two-point reduction of dimensionality. The frequency is calculated for all pairwise combinations and for all possible linkage phase configurations.

cache_counts_twopt(
  input.seq,
  cached = FALSE,
  cache.prev = NULL,
  ncpus = 1L,
  verbose = TRUE,
  joint.prob = FALSE
)

Arguments

input.seq

an object of class mappoly.sequence

cached

If TRUE, access the counts for all linkage phase configurations in a internal file (default = FALSE)

cache.prev

an object of class cache.info containing pre-computed genotype frequencies, obtained with cache_counts_twopt (optional, default = NULL)

ncpus

Number of parallel processes to spawn (default = 1)

verbose

If TRUE (default), print the linkage phase configurations. If cached = TRUE, nothing is printed, since all linkage phase configurations will be cached.

joint.prob

If FALSE (default), returns the frequency of genotypes for transition probabilities (conditional probabilities). If TRUE returns the frequency for joint probabilities. The latter is especially important to compute the Fisher's Information for a pair of markers.

Value

An object of class cache.info which contains one (conditional probabilities) or two (both conditional and joint probabilities) lists. Each list contains all pairs of dosages between parents for all markers in the sequence. The names in each list are of the form 'A-B-C-D', where: A represents the dosage in parent 1, marker k; B represents the dosage in parent 1, marker k+1; C represents the dosage in parent 2, marker k; and D represents the dosage in parent 2, marker k+1. For each list, the frequencies were computed for all possible linkage phase configurations. The frequencies for each linkage phase configuration are distributed in matrices whose names represents the number of homologous chromosomes that share alleles. The rows on these matrices represents the dosages in markers k and k+1 for an individual in the offspring. See Table 3 of S3 Appendix in Mollinari and Garcia (2019) for an example.

References

Mollinari, M., and Garcia, A. A. F. (2019) Linkage analysis and haplotype phasing in experimental autopolyploid populations with high ploidy level using hidden Markov models, _G3: Genes, Genomes, Genetics_. doi:10.1534/g3.119.400378

Author

Marcelo Mollinari, mmollin@ncsu.edu with updates by Gabriel Gesteira, gdesiqu@ncsu.edu

Examples

    all.mrk <- make_seq_mappoly(tetra.solcap, 1:20)
    ## local computation
    counts <- cache_counts_twopt(all.mrk, ncpus = 1)
#> 
#>    Caching the following dosage combination: 
#>         P.k P.k+1 Q.k Q.k+1
#> Conf.1    0     0   1     1
#> Conf.2    0     1   1     0
#> Conf.3    0     1   1     1
#> Conf.4    0     1   1     3
#> Conf.5    0     2   1     1
#> Conf.6    0     3   1     2
#> Conf.7    0     3   1     3
#> Conf.8    0     3   1     4
#> Conf.9    0     4   1     3
#> Conf.10   1     0   0     1
#> Conf.11   1     0   1     1
#> Conf.12   1     1   0     0
#> Conf.13   1     1   0     3
#> Conf.14   1     1   1     0
#> Conf.15   1     1   1     1
#> Conf.16   1     1   1     3
#> Conf.17   1     2   0     1
#> Conf.18   1     2   1     1
#> Conf.19   1     2   3     1
#> Conf.20   1     3   0     2
#> Conf.21   1     3   0     3
#> Conf.22   1     3   0     4
#> Conf.23   1     3   1     2
#> Conf.24   1     3   1     3
#> Conf.25   1     3   1     4
#> Conf.26   1     4   0     3
#> Conf.27   1     4   1     3
#> Conf.28   3     0   2     1
#> Conf.29   3     0   4     1
#> Conf.30   3     1   2     0
#> Conf.31   3     1   2     1
#> Conf.32   3     1   2     3
#> Conf.33   3     1   3     3
#> Conf.34   3     1   4     0
#> Conf.35   3     1   4     1
#> Conf.36   3     1   4     3
#> Conf.37   3     2   2     1
#> Conf.38   3     2   3     1
#> Conf.39   3     2   4     1
#> Conf.40   3     3   2     2
#> Conf.41   3     3   2     3
#> Conf.42   3     3   2     4
#> Conf.43   3     3   4     2
#> Conf.44   3     3   4     3
#> Conf.45   3     3   4     4
#> Conf.46   3     4   2     3
#> Conf.47   3     4   4     3
#> Conf.48   4     0   3     1
#> Conf.49   4     1   3     0
#> Conf.50   4     1   3     1
#> Conf.51   4     1   3     3
#> Conf.52   4     2   3     1
#> Conf.53   4     3   3     2
#> Conf.54   4     3   3     3
#> Conf.55   4     3   3     4
#> Conf.56   4     4   3     3
#> INFO: Going singlemode. Using one Core/CPU/PC for calculation.
#> INFO: Done with 56 phase configurations
#> INFO: Calculation took: 1.648 seconds
    ## load from internal file or web-stored counts (especially important for high ploidy levels)
    counts.cached <- cache_counts_twopt(all.mrk, cached = TRUE)