Concatenate new marker

Inserts a new marker at the end of the sequence, taking into account the two-point information

concatenate_new_marker(X = NULL, d, sh = NULL, seq.num = NULL, ploidy, mrk = 1)

Arguments

X

a list of matrices whose columns represent homologous chromosomes and the rows represent markers

d

the dosage of the inserted marker

sh

a list of shared alleles between all markers in the sequence

seq.num

a vector of integers containing the number of each marker in the raw data file

ploidy

the ploidy level

mrk

the marker to be inserted

Value

a unique list of matrices representing linkage phases