Returns the class with the highest probability in a genotype probability distribution

Returns the class with the highest probability in a genotype probability distribution

dist_prob_to_class(geno, prob.thres = 0.9)

Arguments

geno

the probabilistic genotypes contained in the object 'mappoly.data'

prob.thres

probability threshold to select the genotype. Values below this genotype are assumed as missing data

Value

a matrix containing the doses of each genotype and marker. Markers are disposed in rows and individuals are disposed in columns. Missing data are represented by NAs

Examples

# \donttest{
geno.dose <- dist_prob_to_class(hexafake.geno.dist$geno)
geno.dose$geno.dose[1:10,1:10]
#>      Ind_1 Ind_10 Ind_100 Ind_101 Ind_102 Ind_103 Ind_104 Ind_105 Ind_106
#> M_1      1      1       1       2       1       3       2       2       2
#> M_2      0      1       1       1       0       0       0       0       0
#> M_3      2      2       1       1       1       0       0       0       1
#> M_4      1      1       1       1       1      NA       2       1       0
#> M_5      0      1       0       1       1       1       1       1       0
#> M_6      0      1       0       1       1       1       1       1       0
#> M_7      2      1       1       2       0       0       1       1       1
#> M_8      2      3       3       3       3       4       2       3       3
#> M_9      1      0       0       1       0      NA       1       1       1
#> M_10     3      1       3       2       0       2       2       1       1
#>      Ind_107
#> M_1        2
#> M_2        1
#> M_3        2
#> M_4        1
#> M_5        1
#> M_6        1
#> M_7        2
#> M_8        2
#> M_9        0
#> M_10       3
# }