Filter aneuploid chromosomes from progeny individuals

Filter aneuploid chromosomes from progeny individuals

filter_aneuploid(input.data, aneuploid.info, ploidy)

Arguments

input.data

name of input object (class mappoly.data)

aneuploid.info

data.frame with ploidy information by chromosome (columns) for each individual in progeny (rows). The chromosome and individuals names must match the ones in the file used as input in mappoly.

ploidy

main ploidy

Value

object of class mappoly.data

Author

Cristiane Taniguti, chtaniguti@tamu.edu

Examples

     aneuploid.info <- matrix(4, nrow=tetra.solcap$n.ind, ncol = 12)
     set.seed(8080)
     aneuploid.info[sample(1:length(aneuploid.info), round((4*length(aneuploid.info))/100),0)] <- 3
     aneuploid.info[sample(1:length(aneuploid.info), round((4*length(aneuploid.info))/100),0)] <- 5

     colnames(aneuploid.info) <- paste0(1:12)
     aneuploid.info <- cbind(inds = tetra.solcap$ind.names, aneuploid.info)

     filt.dat <- filter_aneuploid(input.data = tetra.solcap, 
     aneuploid.info = aneuploid.info, ploidy = 4)
#> 8.02 % of the chromosomes x individuals are aneuploids