Import data from polymapR

Function to import datasets from polymapR.

import_data_from_polymapR(
  input.data,
  ploidy,
  parent1 = "P1",
  parent2 = "P2",
  input.type = c("discrete", "probabilistic"),
  prob.thres = 0.95,
  pardose = NULL,
  offspring = NULL,
  filter.non.conforming = TRUE,
  verbose = TRUE
)

Arguments

input.data

a polymapR dataset

ploidy

the ploidy level

parent1

a character string containing the name (or pattern of genotype IDs) of parent 1

parent2

a character string containing the name (or pattern of genotype IDs) of parent 2

input.type

Indicates whether the input is discrete ("disc") or probabilistic ("prob")

prob.thres

threshold probability to assign a dosage to offspring. If the probability is smaller than thresh.parent.geno, the data point is converted to 'NA'.

pardose

matrix of dimensions (n.mrk x 3) containing the name of the markers in the first column, and the dosage of parents 1 and 2 in columns 2 and 3. (see polymapR vignette)

offspring

a character string containing the name (or pattern of genotype IDs) of the offspring individuals. If NULL (default) it considers all individuals as offsprings, except parent1 and parent2.

filter.non.conforming

if TRUE exclude samples with non expected genotypes under no double reduction. Since markers were already filtered in polymapR, the default is FALSE.

verbose

if TRUE (default), the current progress is shown; if FALSE, no output is produced

Details

See examples at https://rpubs.com/mmollin/tetra_mappoly_vignette.

References

Bourke PM et al: (2019) PolymapR — linkage analysis and genetic map construction from F1 populations of outcrossing polyploids. _Bioinformatics_ 34:3496–3502. doi:10.1093/bioinformatics/bty1002

Mollinari, M., and Garcia, A. A. F. (2019) Linkage analysis and haplotype phasing in experimental autopolyploid populations with high ploidy level using hidden Markov models, _G3: Genes, Genomes, Genetics_. doi:10.1534/g3.119.400378

Author

Marcelo Mollinari mmollin@ncsu.edu