Multipoint log-likelihood computation

Update the multipoint log-likelihood of a given map using the method proposed by Mollinari and Garcia (2019).

loglike_hmm(input.map, input.data = NULL, verbose = FALSE)

Arguments

input.map

An object of class mappoly.map

input.data

An object of class mappoly.data, which was used to generate input.map

verbose

If TRUE, map information is shown; if FALSE(default), no output is produced

References

Mollinari, M., and Garcia, A. A. F. (2019) Linkage analysis and haplotype phasing in experimental autopolyploid populations with high ploidy level using hidden Markov models, _G3: Genes, Genomes, Genetics_. doi:10.1534/g3.119.400378

Author

Marcelo Mollinari, mmollin@ncsu.edu

Examples

 # \donttest{
  hexa.map <- loglike_hmm(maps.hexafake[[1]])
  hexa.map
#> This is an object of class 'mappoly.map'
#>     Ploidy level:	 6 
#>     No. individuals:	 300 
#>     No. markers:	 538 
#>     No. linkage phases:	 1 
#> 
#>     ---------------------------------------------
#>     Number of linkage phase configurations:  1
#>     ---------------------------------------------
#>     Linkage phase configuration:  1
#>        map length:	 116.39
#>        log-likelihood:	 -6135.65
#>        LOD:		 0
#>     ~~~~~~~~~~~~~~~~~~
 # }