Plots mappoly.prefpair.profiles

Plots mappoly.prefpair.profiles

# S3 method for mappoly.prefpair.profiles
plot(
  x,
  type = c("pair.configs", "hom.pairs"),
  min.y.prof = 0,
  max.y.prof = 1,
  thresh = 0.01,
  P1 = "P1",
  P2 = "P2",
  ...
)

Arguments

x

an object of class mappoly.prefpair.profiles

type

a character string indicating which type of graphic is plotted: "pair.configs" (default) plots the preferential pairing profile for the pairing configurations or "hom.pairs" plots the preferential pairing profile for the homolog pairs

min.y.prof

lower bound for y axis on the probability profile graphic (default = 0)

max.y.prof

upper bound for y axis on the probability profile graphic (default = 1)

thresh

threshold for chi-square test (default = 0.01)

P1

a string containing the name of parent P1

P2

a string containing the name of parent P2

...

unused arguments