Autotetraploid potato dataset.

A dataset of the B2721 population which derived from a cross between two tetraploid potato varieties: Atlantic × B1829-5. The population comprises 160 offsprings genotyped with the SolCAP Infinium 8303 potato array. The original data set can be found in [The Solanaceae Coordinated Agricultural Project (SolCAP) webpage](http://solcap.msu.edu/potato_infinium.shtml) The dataset also contains the genomic order of the SNPs from the Solanum tuberosum genome version 4.03. The genotype calling was performed using the fitPoly R package.

tetra.solcap

Format

An object of class mappoly.data which contains a list with the following components:

ploidy

ploidy level = 4

n.ind

number individuals = 160

n.mrk

total number of markers = 4017

ind.names

the names of the individuals

mrk.names

the names of the markers

dosage.p1

a vector containing the dosage in parent P for all n.mrk markers

dosage.p2

a vector containing the dosage in parent Q for all n.mrk markers

chrom

a vector indicating the chromosome each marker belongs. Zero indicates that the marker was not assigned to any sequence

genome.pos

Physical position of the markers into the sequence

geno.dose

a matrix containing the dosage for each markers (rows) for each individual (columns). Missing data are represented by ploidy_level + 1 = 5

n.phen

There are no phenotypes in this simulation

phen

There are no phenotypes in this simulation

chisq.pval

vector containing p-values for all markers associated to the chi-square test for the expected segregation patterns under Mendelian segregation