Add markers that are informative in both parents using HMM approach and evaluating difference in LOD and gap size

Add markers that are informative in both parents using HMM approach and evaluating difference in LOD and gap size

update_framework_map(
  input.map.list,
  input.seq,
  twopt,
  thres.twopt = 10,
  init.LOD = 30,
  verbose = TRUE,
  method = "hmm",
  input.mds = NULL,
  max.rounds = 50,
  size.rem.cluster = 2,
  gap.threshold = 4
)

Arguments

input.map.list

list containing three mappoly.map objects:1) map built with markers with segregation information from parent 1; 2) map built with markers with segregation information from parent 2; 3) maps in 1 and 2 merged

input.seq

object of class mappoly.sequence containing all markers for specific group

twopt

object of class mappoly.twopt

thres.twopt

the LOD threshold used to determine if the linkage phases compared via two-point analysis should be considered for the search space reduction (A.K.A. η in Mollinari and Garcia (2019), default = 5)

init.LOD

the LOD threshold used to determine if the marker will be included or not after hmm analysis (default = 30)

verbose

If TRUE (default), current progress is shown; if FALSE, no output is produced

method

indicates whether to use 'hmm' (Hidden Markov Models), 'ols' (Ordinary Least Squares) or 'wMDS_to_1D_pc' (weighted MDS followed by fitting a one dimensional principal curve) to re-estimate the recombination fractions after adding markers

input.mds

An object of class mappoly.map

max.rounds

integer defining number of times to try to fit the remaining markers in the sequence

size.rem.cluster

threshold for number of markers that must contain in a segment after a gap is removed to keep this segment in the sequence

gap.threshold

threshold for gap size

Value

object of class mappoly.map2

Author

Marcelo Mollinari, mmollin@ncsu.edu with documentation and minor modifications by Cristiane Taniguti chtaniguti@tamu.edu