Update map

This function takes an object of class mappoly.map and checks for removed redundant markers in the original dataset. Once redundant markers are found, they are re-added to the map in their respective equivalent positions and another HMM round is performed.

update_map(input.maps, verbose = TRUE)

Arguments

input.maps

a single map or a list of maps of class mappoly.map

verbose

if TRUE (default), shows information about each update process

Value

an updated map (or list of maps) of class mappoly.map, containing the original map(s) plus redundant markers

Author

Gabriel Gesteira, gdesiqu@ncsu.edu

Examples

orig.map <- solcap.err.map
up.map <- lapply(solcap.err.map, update_map)
#> Updating map 1 
#> Updating map 1 
#> Updating map 1 
#> Updating map 1 
#> Updating map 1 
#> Updating map 1 
#> Updating map 1 
#> Updating map 1 
#> Updating map 1 
#> Updating map 1 
#> Updating map 1 
#> Updating map 1 
summary_maps(orig.map)
#> 
#> Your dataset contains removed (redundant) markers. Once finished the maps, remember to add them back with the function 'update_map'.
#>       LG Genomic sequence Map length (cM) Markers/cM Simplex Double-simplex
#> 1      1                1          316.43       1.18     108             24
#> 2      2                2          202.85       1.23      88             49
#> 3      3                3          102.26       2.66     104             47
#> 4      4                4          101.93       3.52     143             34
#> 5      5                5           82.98       2.95      75             22
#> 6      6                6          181.58       2.02     100             55
#> 7      7                7          134.58       2.44     110             55
#> 8      8                8          117.42       2.14      71             36
#> 9      9                9          202.32       1.32      85             46
#> 10    10               10          127.32       1.55      70             31
#> 11    11               11          101.28       2.34      79             35
#> 12    12               12          162.09       1.21      61             35
#> 13 Total             <NA>         1833.04       2.05    1094            469
#>    Multiplex Total Max gap
#> 1        241   373    9.69
#> 2        112   249   10.16
#> 3        121   272    6.54
#> 4        182   359    3.07
#> 5        148   245    4.43
#> 6        212   367    3.94
#> 7        163   328    9.81
#> 8        144   251   10.77
#> 9        136   267   11.07
#> 10        96   197    7.15
#> 11       123   237    6.32
#> 12       100   196    6.37
#> 13      1778  3341    7.44
summary_maps(up.map)
#>       LG Genomic sequence Map length (cM) Markers/cM Simplex Double-simplex
#> 1      1                1          316.44       1.24     108             24
#> 2      2             2-NA          202.85       1.29      88             49
#> 3      3                3          102.26       2.93     104             47
#> 4      4                4          101.93       3.91     143             34
#> 5      5                5           82.98       3.09      75             22
#> 6      6             6-NA          181.59       2.24     100             55
#> 7      7             7-NA          134.59       2.79     110             55
#> 8      8                8          117.42       2.25      71             36
#> 9      9                9          202.32       1.39      85             46
#> 10    10               10          127.32        1.7      70             31
#> 11    11               11          101.29       2.66      79             35
#> 12    12               12          162.09       1.25      61             35
#> 13 Total             <NA>         1833.08       2.23    1094            469
#>    Multiplex Total Max gap
#> 1        241   393    9.69
#> 2        112   262   10.16
#> 3        121   300    6.54
#> 4        182   399    3.07
#> 5        148   256    4.43
#> 6        212   406    3.94
#> 7        163   375    9.81
#> 8        144   264   10.77
#> 9        136   281   11.07
#> 10        96   216    7.15
#> 11       123   269    6.32
#> 12       100   202    6.37
#> 13      1778  3623    7.44